Create yellowPageScraper.py
Program that scrapes the Yellow Pages website for businesses that do not have websites. Presently, program users go to the Yellow Pages, type in their search query, and copy paste the url from the subsequent page(s) containing the results to urls.txt. The scraper iterates through each URL in the text file and writes the business name, email, phone number, and address from any matches to text and csv/excel files. 15-20 minutes of searching the permutations of AA-ZZ for a city can result in thousands of potential business data points being scraped. This is a great tool for freelancers looking for clients. In time, I will automate the permutation search and the process of gathering data for subsequent pages in each permutation so users will no longer have to copy and paste the search urls to the urls.txt file. This is already MUCH more efficient than doing this by hand though.